% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swear.R
\name{swear}
\alias{swear}
\title{Swear}
\usage{
swear(method = getOption("swear.method", "explicit"), n = 1L, id = NULL)
}
\arguments{
\item{method}{A method of swearing}

\item{n}{The number of swears}

\item{id}{The integer identification of a curse}
}
\description{
Say some sh*t
}
