% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swear.R
\docType{data}
\name{swears}
\alias{swears}
\title{Swears}
\format{
A data frame with xx rows and 2 variables
\describe{
\item{explicit}{Swearing phrases}
\item{alternative}{Alternative phrases}
}
}
\usage{
swears
}
\description{
A data frame of swears with the explicit and alternative swears.
}
\keyword{datasets}
